Add decorator classes
