--- person.rb
@@ -1,27 +1,0 @@
-class Person
-  def initialize(age, name = 'Unkown', parent_permission: true)
-    @id = Random.rand(1...1000)
-    @name = name
-    @age = age
-    @parent_permission = parent_permission
-  end
-
-  attr_accessor :name, :age
-  attr_reader :id
-
-  private
-
-  def of_age?
-    @age >= 18
-  end
-
-  public
-
-  def can_use_services?
-    if of_age? || @parent_permission
-      true
-    else
-      false
-    end
-  end
-end
