+++ src/baseDecorator.rb
@@ -1,0 +1,12 @@
+require './src/nameable'
+
+class Decorator < Nameable
+  def initialize(nameable)
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
