+++ src/capitalize.rb
@@ -1,0 +1,7 @@
+require './src/baseDecorator'
+
+class Capitalizer < Decorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
