+++ src/trimmer.rb
@@ -1,0 +1,6 @@
+require './src/baseDecorator'
+class Trimmer < Decorator
+  def correct_name
+    @nameable.correct_name[0..9]
+  end
+end
